Move deep learning with python examples to separate directory
