+++ src/CntkCatalyst.Examples/CntkCatalyst.Examples.csproj
@@ -80,13 +80,13 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Ch_21_First_Look_At_A_Neural_Network.cs" />
-    <Compile Include="Ch_62_Understanding_Recurrent_Neural_Networks.cs" />
-    <Compile Include="Ch_64_Sequence_Processing_With_Convnets.cs" />
-    <Compile Include="Ch_61_Using_Word_Embeddings.cs" />
-    <Compile Include="Ch_51_Introduction_To_Convnets.cs" />
-    <Compile Include="Ch_35_Classifying_Movie_Reviews.cs" />
-    <Compile Include="Ch_52_Using_Convnets_With_Small_Datasets.cs" />
+    <Compile Include="DeepLearningFrancoisChollet\Ch_21_First_Look_At_A_Neural_Network.cs" />
+    <Compile Include="DeepLearningFrancoisChollet\Ch_62_Understanding_Recurrent_Neural_Networks.cs" />
+    <Compile Include="DeepLearningFrancoisChollet\Ch_64_Sequence_Processing_With_Convnets.cs" />
+    <Compile Include="DeepLearningFrancoisChollet\Ch_61_Using_Word_Embeddings.cs" />
+    <Compile Include="DeepLearningFrancoisChollet\Ch_51_Introduction_To_Convnets.cs" />
+    <Compile Include="DeepLearningFrancoisChollet\Ch_35_Classifying_Movie_Reviews.cs" />
+    <Compile Include="DeepLearningFrancoisChollet\Ch_52_Using_Convnets_With_Small_Datasets.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
