+++ src/CntkCatalyst.Examples/DeepLearningFrancoisChollet/Ch_21_First_Look_At_A_Neural_Network.cs
@@ -7,7 +7,7 @@ using CNTK;
 using CntkCatalyst.LayerFunctions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace CntkCatalyst.Examples
+namespace CntkCatalyst.Examples.DeepLearningFrancoisChollet
 {
     /// <summary>
     /// Example from Chapter 02: First look at a Neural Network:
