+++ src/CntkCatalyst.Examples/DeepLearningFrancoisChollet/Ch_52_Using_Convnets_With_Small_Datasets.cs
@@ -6,7 +6,7 @@ using CNTK;
 using CntkCatalyst.LayerFunctions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace CntkCatalyst.Examples
+namespace CntkCatalyst.Examples.DeepLearningFrancoisChollet
 {
     /// <summary>
     /// Example from Chapter 5.2: Using convnets with small datasets:
