+++ src/CntkCatalyst.Examples/DeepLearningFrancoisChollet/Ch_61_Using_Word_Embeddings.cs
@@ -7,12 +7,11 @@ using CNTK;
 using CntkCatalyst.LayerFunctions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace CntkCatalyst.Examples
+namespace CntkCatalyst.Examples.DeepLearningFrancoisChollet
 {
     /// <summary>
-    /// Example from Chapter 3.5: Classifying movie reviews:
-    /// https://github.com/mdabros/deep-learning-with-python-notebooks/blob/master/3.5-classifying-movie-reviews.ipynb
-    ///
+    /// Example from Chapter 6.1: Using word embeddings
+    /// https://github.com/mdabros/deep-learning-with-python-notebooks/blob/master/6.1-using-word-embeddings.ipynb
     ///
     /// This example needs manual download of the IMDB dataset in sparse CNTK format.
     /// The data can found following the link below, 
@@ -20,7 +19,7 @@ namespace CntkCatalyst.Examples
     /// https://msdn.microsoft.com/en-us/magazine/mt830362
     /// </summary>
     [TestClass]
-    public class Ch_35_Classifying_Movie_Reviews
+    public class Ch_61_Using_Word_Embeddings
     {
         [TestMethod]
         public void Run()
@@ -45,11 +44,9 @@ namespace CntkCatalyst.Examples
             var biasInit = Initializers.Zero();
 
             // Create the architecture.
-            //var input = CNTKLib.SequenceLast(Layers.Input(inputShape, dataType));
             var network = Layers.Input(inputShape, dataType, isSparse: true)
-                .Dense(16, weightInit(), biasInit, device, dataType)
-                .ReLU()
-                .Dense(16, weightInit(), biasInit, device, dataType)
+                .Embedding(8, weightInit(), dataType, device)
+                .Dense(32, weightInit(), biasInit, device, dataType)
                 .ReLU()
                 .Dense(numberOfClasses, weightInit(), biasInit, device, dataType)
                 .Softmax();
@@ -64,10 +61,10 @@ namespace CntkCatalyst.Examples
                 dynamicAxes: new List<Axis>() { Axis.DefaultBatchAxis() },
                 isSparse: false);
 
-            // setup loss and metric.
+            // setup loss and learner.
             var lossFunc = Losses.CategoricalCrossEntropy(network.Output, targetVariable);
             var metricFunc = Metrics.Accuracy(network.Output, targetVariable);
-            
+
             // setup trainer.
             var learner = Learners.Adam(network.Parameters());
             var trainer = Trainer.CreateTrainer(network, lossFunc, metricFunc, new List<Learner> { learner });
@@ -102,7 +99,7 @@ namespace CntkCatalyst.Examples
             var testSource = new CntkMinibatchSource(test, nameToVariable);
 
             // Train the model using the training set.
-            var history = model.Fit(trainingSource, epochs: 20, batchSize: 512,
+            var history = model.Fit(trainingSource, epochs: 100, batchSize: 512,
                 validationMinibatchSource: testSource);
 
             // Trace loss and validation history
